feat: migrate auto test task page to dynamic routing with server-side data fetching
